// src/api-chain/model/chainDocumentRequirementList.ts
export interface ChainDocumentRequirementList { 
    /**
     * Identifier for document requirement (defining type)
     */
    identifier: string;
    /**
     * Semi product for which document requirement list is relevant.
     */
    semiProductId?: string;
    /**
     * List of document requirements
     */
    requirements: Array<ChainDocumentRequirement>;
    targets: ScoreTargets;
}

/**
 * Namespace for property- and property-value-enumerations of ChainDocumentRequirementList.
 */
export namespace ChainDocumentRequirementList {
    /**
     * All properties of ChainDocumentRequirementList.
     */
    export enum Properties {
        /**
         * Identifier for document requirement (defining type)
         */
        identifier = 'identifier',
        /**
         * Semi product for which document requirement list is relevant.
         */
        semiProductId = 'semiProductId',
        /**
         * List of document requirements
         */
        requirements = 'requirements',
        targets = 'targets'
    }


    export function formMetadata() {
        return  {
            metadata: formMetadata,
            classname: 'ChainDocumentRequirementList',
            vars: [
                {
                    isReadOnly: false,
                    isEnum: false,
                    required: true,
                    name: 'identifier',
                    classname: 'ChainDocumentRequirementList',
                    dataType: 'string',
                    isPrimitiveType: true,
                    isListContainer: false,
                    complexType: ''
                },
                {
                    isReadOnly: false,
                    isEnum: false,
                    required: false,
                    name: 'semiProductId',
                    classname: 'ChainDocumentRequirementList',
                    dataType: 'string',
                    isPrimitiveType: true,
                    isListContainer: false,
                    complexType: ''
                },
                {
                    metadata: ChainDocumentRequirement.formMetadata,
                    isReadOnly: false,
                    isEnum: false,
                    required: true,
                    name: 'requirements',
                    classname: 'ChainDocumentRequirementList',
                    dataType: 'Array&lt;ChainDocumentRequirement&gt;',
                    isPrimitiveType: false,
                    isListContainer: true,
                    complexType: 'ChainDocumentRequirement'
                },
                {
                    metadata: ScoreTargets.formMetadata,
                    isReadOnly: false,
                    isEnum: false,
                    required: true,
                    name: 'targets',
                    classname: 'ChainDocumentRequirementList',
                    dataType: 'ScoreTargets',
                    isPrimitiveType: false,
                    isListContainer: false,
                    complexType: 'ScoreTargets'
                },
            ],
            validators: {
                identifier: [
                        ['required'],
                ],
                semiProductId: [
                ],
                requirements: [
                        ['required'],
                ],
                targets: [
                        ['required'],
                ],
            }
        }
    }

  // export const ChainDocumentRequirementListValidationScheme = {
  //     validators: [],
  //     fields: {
  //               identifier: {
  //                   validators: []
  //               },
  //               semiProductId: {
  //                   validators: []
  //               },
  //               requirements: {
  //                   validators: []
  //               },
  //               targets: {
  //                   validators: []
  //               },
  //     }
  // } as SimpleValidationScheme<ChainDocumentRequirementList>;


}